Remove data insertion from migration scripts
We didn't want to include secrets like phone numbers in the migration script, and really that didn't make
sense with the idea of migrations being schema changes. So instead split up the 002 script into one that
created the installation table and references to it without constraints, and a 003 script that would add
the constraints after a user goes in and manually inserts the data.