Debugging Travis
Travis CI is causing DB connection issues, where there weren't before.
You can see this by re-running the jobs for previously passing and
already merging commits on the main branch.

I am going to investigate!

Conclusion: It seems Travis CI's instance of PostgreSQL is now running
on port 5432, and not 5433. That's annoying, but it seems to be working
now.